Move cli/util cloud based stuff into a cli/cloud package.
